Move social buttons to the personal settings
