--- settings/templates/settings/admin/server.development.notice.php
@@ -1,3 +1,0 @@
-<div class="section">
-	<p><?php include __DIR__ . '/../../settings.development.notice.php'; ?></p>
-</div>
